requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==22.1
python-decouple==3.6
psycopg2-binary==2.9.1
sentry-sdk==1.17.0
django-crontab==0.7.1
Faker==18.3.1